Move all Python 2/3 compatibility code in compat.py
