+++ edn_format/char.py
@@ -1,9 +1,6 @@
 # -*- coding: utf-8 -*-
 
-import sys
-
-if sys.version_info[0] == 3:
-    unicode = str
+from .compat import unicode
 
 
 class Char(unicode):
